project2/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pika==1.3.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3.post1
sendgrid==6.10.0
six==1.16.0
starkbank-ecdsa==2.2.0
Werkzeug==3.0.1
